Detect std::bit_cast only in C++20 mode
