+++ clang-tools-extra/test/clang-tidy/checkers/bugprone/bitwise-pointer-cast.cpp
@@ -7,32 +7,10 @@ void memcpy(void* to, void* dst, unsigned long long size)
 
 namespace std
 {
-template <typename To, typename From>
-To bit_cast(From from)
-{
-  // Dummy implementation for the purpose of the test
-  To to{};
-  return to;
-}
-
 using ::memcpy;
 }
 
 void pointer2pointer()
-{
-  int x{};
-  float bad = *std::bit_cast<float*>(&x); // UB, but looks safe due to std::bit_cast
-  // CHECK-MESSAGES: :[[@LINE-1]]:16: warning: do not use 'std::bit_cast' to cast between pointers [bugprone-bitwise-pointer-cast]
-  float good = std::bit_cast<float>(x);   // Well-defined
-
-  using IntPtr = int*;
-  using FloatPtr = float*;
-  IntPtr x2{};
-  float bad2 = *std::bit_cast<FloatPtr>(x2);
-  // CHECK-MESSAGES: :[[@LINE-1]]:17: warning: do not use 'std::bit_cast' to cast between pointers [bugprone-bitwise-pointer-cast]
-}
-
-void pointer2pointer_memcpy()
 {
   int x{};
   int* px{};
@@ -51,11 +29,13 @@ void pointer2pointer_memcpy()
 void int2pointer()
 {
   unsigned long long addr{};
-  float* p = std::bit_cast<float*>(addr);
+  float* p{};
+  std::memcpy(&p, &addr, sizeof(addr));
 }
 
 void pointer2int()
 {
+  unsigned long long addr{};
   float* p{};
-  auto addr = std::bit_cast<unsigned long long>(p);
+  std::memcpy(&addr, &p, sizeof(p));
 }
